Keep the lens demo's expect/actuals with the demo
The shared settings files go back to their original scope; the lens
demo's render-mode section declares its expects in the demo file and its
actuals under demos/ in each platform source set, like OfflineRegionSection.
