+++ demo-app/common/src/commonMain/kotlin/org/maplibre/compose/demoapp/DemoSettings.kt
@@ -31,23 +31,6 @@ class DemoSettings {
  */
 @Composable expect fun RenderSettingsItems(settings: DemoSettings)
 
-/**
- * The magnifying lens demo's rendering rows: the app's render mode and the lens map's own, on
- * platforms that offer the choice. Empty on the web target, which has no render mode.
- */
-@Composable
-expect fun LensRenderSection(
-  settings: DemoSettings,
-  lensOptions: RenderOptions,
-  onLensChange: (RenderOptions) -> Unit,
-)
-
-/**
- * The lens map's initial render options: texture mode where a texture-versus-surface choice exists,
- * because Android applies Compose modifiers to the map only in texture mode.
- */
-expect val LensRenderOptionsDefault: RenderOptions
-
 /** Presets for [TileLodOptions], shared because every platform exposes the same three. */
 @Composable
 fun TileLodSettingsItems(settings: DemoSettings) {
