+++ demo-app/common/src/jsMain/kotlin/org/maplibre/compose/demoapp/RenderSettingsItems.js.kt
@@ -3,7 +3,6 @@ package org.maplibre.compose.demoapp
 import androidx.compose.runtime.Composable
 import org.maplibre.compose.demoapp.design.FpsCapRow
 import org.maplibre.compose.demoapp.design.SwitchRow
-import org.maplibre.compose.map.RenderOptions
 
 @Composable
 actual fun RenderSettingsItems(settings: DemoSettings) {
@@ -22,12 +21,3 @@ actual fun RenderSettingsItems(settings: DemoSettings) {
     settings.renderOptions = options.copy(isOverdrawInspectorEnabled = it)
   }
 }
-
-@Composable
-actual fun LensRenderSection(
-  settings: DemoSettings,
-  lensOptions: RenderOptions,
-  onLensChange: (RenderOptions) -> Unit,
-) {}
-
-actual val LensRenderOptionsDefault: RenderOptions = RenderOptions.Standard
